Finished initial port to Arduboy
Now the game successfully runs on the Arduboy emulator, but the
graphics are corrupted and the screen resolution is wrong.  Also,
all the text in the menus isn't aligned right, as it was originally
designed with a smaller font.  However, it does appear to run well,
so I think all that needs to be done is fixing all the display
issues and it will work.  There's still 5% PROGMEM left, so
technically there might be room to add something.  I had to
comment out the gamebuino sound library calls, so there is no
sound now, but I don't know if 5% is enough to add back sound.