Fixed sprite display bug
I knew that the reason the sprites looked glitched was that the
Gamebuino stores the dimensions of its sprites along with the rest
of the sprite data, while the Arduboy requires that information to
be provided when the draw functions are called.  Therefore I needed
to delete all the dimensions from the sprite data and then change
the offset calculations, and now it looks right!  I also inverted
the screen so that white is the background color, since that's
how I designed the graphics on the Gamebuino.  Theoretically having
more white on the screen could increase battery drain, but I doubt
that's a huge problem.