Fixed bugs caused by expanding the viewport
The world map was designed with the screen size of the Gamebuino
in mind, so that the edge of the map would just barely go offscreen
and give the appearance that there was more city just beyond the
edge of where the player was.  Unfortunately, by expanding the area
drawn around the player, I had some drawing glitches.  First,
underflow of the calculation of the x coordinate meant the world
map drawing would go blank, and after fixing that it meant that
the edge of the world map cut off immediately in an ugly way.
I decided to fix this by extending edge tiles on the left, and
adding water tiles on the right.  This mostly works well, although
there are minor artifacts on both the left and right sides.

I also fixed a couple of other minor display issues, such as the
health bar not displaying and some text overdrawing the border of
the text box.