Changed EEPROM save base, fixed text formatting
I think settings are saved at the low bytes of EEPROM, so I changed
the location the game data is saved to.

I also fixed all the broken text formatting to correspond to the
size of the default Arduboy font, which is a bit larger than the
default Gamebuino font.