Added text on pause screen to indicate current objective
I felt like the game doesn't provide enough guidance to a player,
and if someone steps away from the game and comes back, they might
not know what they need to do next.  I tried to help with that by
giving a short description of the current objective in the main
quest.

I'm also considering showing which floor of a dungeon the player
is on, and I need to fix a bug where hitting the menu button during
dialogue breaks the game.