Fixed save bug, shrunk collision box, changed some color and menus
When I made this game for the Gamebuino, I just saved at the start
of eeprom.  However, the first 16 bytes are reserved on the
Arduboy, so I shifted it to a different base address.
Unfortunately, I didn't check my changes closely enough, and most
of the data was saved near the start of eeprom, partially over the
reserved space.  I noticed this because the device name that was
displayed at start up changed as save data changed, which made me
decide to investigate it.

One frustrating thing about the game, originally, was that certain
areas had gaps exactly as wide as the character, so it was hard to
get in certain openings.  Movement in dungeons was especially
frustrating.  I wasn't sure of the best way to improve this, so I
waited and didn't make any changes.  However, with all the small
improvements I was making for this port, I decided to just try
shrinking the player's bounding box by one pixel on all sides.  I
wasn't sure whether it might look strange to partially intersect
the world tiles, but it turned out it looks ok!  I think this makes
movement much more pleasant.  I may go back and add this
improvement to the original game.

I also changed the dialogue to use the same colors as in the
original game, and changed the backgrounds of the menus to be
black.  I removed the unneeded "quit" menu option in the main menu,
since that had been used to go to the Gamebuino loader.