Fixed saving bug and minor dungeon collision bug
It turned out that changing the player's collision had the
unfortunate side-effect of making it possible to save in a location
in which when the game is loaded the player is inside a wall.  I
fixed this by simply adding a small offset to the player's position
before dividing it by 8 and saving it, so it should always save a
safe pair of coordinates now.

I also fixed a minor collision bug that I had noticed in dungeons.
I set one offset wrong, which allowed the player to walk though
walls in very specific circumstances.  It did not appear to let
the player get stuck or go out of bounds, so it was not critical,
but it's good to fix it.